Condense logging headers
